+++ NEWS.md
@@ -4,6 +4,7 @@
 - Creating a service-specific endpoint (`file_endpoint`, `blob_endpoint`, `adls_endpoint`) with an invalid URL will now warn, instead of throwing an error. This enables using tools like Azurite, which use a local address as the endpoint. Calling `storage_endpoint` with an invalid URL will still throw an error, as the function has no way of telling which storage service is required.
 - Remove the warning about ADLSgen2 not supporting shared access signatures (SAS).
 - Background process pool functionality has been moved to AzureRMR. This removes code duplication, and also makes it available for other packages that may benefit.
+- Only display the file transfer progress bar in an interactive session.
 
 # AzureStor 2.1.1